refactor. 리다이렉트 유틸(api→lib) 이동
